Correction bugs

+++ lib/libft/ft_bzero.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rbalazs <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: rbalazs <rbalazs@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 11:03:18 by rbalazs           #+#    #+#             */
-/*   Updated: 2023/12/13 11:03:21 by rbalazs          ###   ########.fr       */
+/*   Updated: 2024/06/21 14:59:02 by rbalazs          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *str, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 	unsigned char	*chaine;
 
 	i = 0;
